Refactored fetch request extensions to challenge into own file
